requirement.txt
# Core dependencies
torch>=1.12.0
torchvision>=0.13.0
pytorch-lightning>=1.7.0
numpy>=1.22.0
scipy>=1.8.0
matplotlib>=3.5.0
pillow>=9.0.0
tqdm>=4.64.0
omegaconf>=2.2.0
pyyaml>=6.0

# Image processing and metrics
opencv-python>=4.6.0
pyiqa>=0.1.5  # Image quality assessment
lpips>=0.1.4
pytorch-fid>=0.3.0

# Data handling and utilities
pandas>=1.4.0
natsort>=8.1.0  # For natural sorting
h5py>=3.7.0
einops>=0.5.0
tensorboard>=2.10.0

# Optional for logging
wandb>=0.13.0

# For Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.15.0

# For development
black>=22.6.0
isort>=5.10.0